Project tree, import methods changed for the creating package. setup.py , _version.py created
