+++ README.md
@@ -5,12 +5,14 @@ for install:
 pip install -r requirements.txt
 ```
 
+
 for run:
 ```
-cd examples
+cd src/examples
 python hello_world.py
 ```
 
 open browser: http://localhost:5000
 
-"asdasf"
+
+
