+++ src/ait_ui/__init__.py
@@ -1,0 +1,3 @@
+from . import _version
+__version__ = _version.__version__
+del _version
