+++ src/ait_ui/_version.py
@@ -1,0 +1,2 @@
+# Master Versino AIT-UI
+__version__ = '0.1.0'
