+++ src/ait_ui/components/col.py
@@ -1,4 +1,4 @@
-from components.element import Element
+from .element import Element
 class Col(Element):
     def __init__(self,id = None, value = None):
         super().__init__(id, value)
