+++ src/ait_ui/components/file.py
@@ -1,6 +1,14 @@
-from components.element import Element
-import components.scripts as scripts
+from .element import Element
+from . import scripts 
+
+#---------------------------------------------------------------#
+## To acces the connection  from up directory we use the following code
+import os 
+import sys
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
 import connection
+#---------------------------------------------------------------#
+
 scripts.header_items.append("<script src='https://cdnjs.cloudflare.com/ajax/libs/dropzone/5.9.3/dropzone.min.js'></script>")
 
 scripts.add_script("dropzone", """
