+++ src/ait_ui/components/imageviewer.py
@@ -1,7 +1,13 @@
-from components.element import Element
-import connection
+from .element import Element
+from . import scripts
 
-import components.scripts as scripts
+#---------------------------------------------------------------#
+## To acces the connection  from up directory we use the following code
+import os
+import sys
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
+import connection
+#---------------------------------------------------------------#
 
 scripts.header_items.append('<script src="https://cdnjs.cloudflare.com/ajax/libs/openseadragon/4.1.0/openseadragon.min.js"></script>')
