+++ src/ait_ui/components/link.py
@@ -1,4 +1,4 @@
-from components.element import Element
+from .element import Element
 
 class Link(Element):
     def __init__(self,id = None,value = None,href = None):
