+++ src/ait_ui/static/style.css
