+++ src/examples/canvas.py
@@ -4,19 +4,19 @@
 # on production does not need to be added
 import sys
 import os
-sys.path.append(os.path.abspath(os.path.join('..')))
-#---------------------------------------------------------------#
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
+#-------------------------------------------------------------
 
-import app
+from ait_ui import app
 
-from components.element import Element, Elm
-from components.text import Text
-from components.image import Image
-from components.imageviewer import ImageViewer
-from components.canvas import Canvas
-from components.row import Row
-from components.button import Button
-from components.slider import Slider
+from ait_ui.components.element import Element, Elm
+from ait_ui.components.text import Text
+from ait_ui.components.image import Image
+from ait_ui.components.imageviewer import ImageViewer
+from ait_ui.components.canvas import Canvas
+from ait_ui.components.row import Row
+from ait_ui.components.button import Button
+from ait_ui.components.slider import Slider
 
 mouse_down = False
 selected_color = "red"
